lot of fixing and refactoring for EDWH-4162
