+++ streamlit_rag/requirements.txt
@@ -6,3 +6,4 @@ pydantic
 pydantic-settings
 pyparsing
 click
+tenacity
